replace trtllm_latency
